backend/requirements.txt
Flask>=2.3.3
Flask-Cors>=4.0.0
python-dotenv>=1.0.0
openai>=1.0.0
openai-agents>=0.1.0
requests>=2.28.0
asyncio>=3.4.3
arize-phoenix>=2.0.0
arize-phoenix-otel>=1.0.0
openinference-instrumentation-openai>=1.0.0
openinference-instrumentation-openai-agents>=1.0.0
opentelemetry-instrumentation-flask
openinference-instrumentation>=1.0.0 